+++ modules/tinymce/Gruntfile.js
@@ -8,7 +8,7 @@ let swag = require('@ephox/swag');
 let plugins = [
   'advlist', 'anchor', 'autolink', 'autoresize', 'autosave', 'charmap', 'code', 'codesample',
   'directionality', 'emoticons', 'help', 'fullscreen', 'image', 'importcss',
-  'insertdatetime', 'link', 'lists', 'media', 'nonbreaking', 'noneditable', 'pagebreak', 'paste',
+  'insertdatetime', 'link', 'lists', 'media', 'nonbreaking', 'pagebreak', 'paste',
   'preview', 'save', 'searchreplace', 'table', 'template',
   'visualblocks', 'visualchars', 'wordcount', 'quickbars'
 ];