+++ modules/tinymce/src/core/main/ts/api/Options.ts
@@ -30,6 +30,8 @@ const getHash = (value: string): Record<string, string> => {
   }, {} as Record<string, string>);
 };
 
+const isRegExp = (x: unknown): x is RegExp => Type.is(x, RegExp);
+
 const option = <K extends keyof EditorOptions>(name: K) => (editor: Editor) =>
   editor.options.get(name);
 
@@ -694,6 +696,29 @@ const register = (editor: Editor) => {
     ]
   });
 
+  registerOption('noneditable_class', {
+    processor: 'string',
+    default: 'mceNonEditable'
+  });
+
+  registerOption('editable_class', {
+    processor: 'string',
+    default: 'mceEditable'
+  });
+
+  registerOption('noneditable_regexp', {
+    processor: (value) => {
+      if (Type.isArrayOf(value, isRegExp)) {
+        return { value, valid: true };
+      } else if (isRegExp(value)) {
+        return { value: [ value ], valid: true };
+      } else {
+        return { valid: false, message: 'Must be a RegExp or an array of RegExp.' };
+      }
+    },
+    default: []
+  });
+
   registerOption('table_clone_elements', {
     processor: 'string[]'
   });
@@ -833,6 +858,9 @@ const shouldBrowserSpellcheck = option('browser_spellcheck');
 const getProtect = option('protect');
 const getContentEditableState = option('content_editable_state');
 const getTextPatterns = option('text_patterns');
+const getNonEditableClass = option('noneditable_class');
+const getEditableClass = option('editable_class');
+const getNonEditableRegExps = option('noneditable_regexp');
 
 const getFontStyleValues = (editor: Editor): string[] =>
   Tools.explode(editor.options.get('font_size_style_values'));
@@ -963,6 +991,9 @@ export {
   getProtect,
   getContentEditableState,
   getTextPatterns,
+  getNonEditableClass,
+  getNonEditableRegExps,
+  getEditableClass,
   hasTableTabNavigation,
   getTableCloneElements,
   hasTableObjectResizing,