+++ modules/tinymce/src/plugins/emoticons/main/ts/core/EmojiDatabase.ts
@@ -47,8 +47,8 @@ const categoryNameMap = {
 const translateCategory = (categories: Record<string, string>, name: string): string =>
   Obj.has(categories, name) ? categories[name] : name;
 
-const getUserDefinedEmoticons = (editor: Editor): Record<string, RawEmojiEntry> => {
-  const userDefinedEmoticons = Options.getAppendedEmoticons(editor);
+const getUserDefinedEmoji = (editor: Editor): Record<string, RawEmojiEntry> => {
+  const userDefinedEmoticons = Options.getAppendedEmoji(editor);
   return Obj.map(userDefinedEmoticons, (value) =>
     // Set some sane defaults for the custom emoji entry
     ({ keywords: [], category: 'user', ...value })
@@ -60,7 +60,7 @@ const initDatabase = (editor: Editor, databaseUrl: string, databaseId: string): 
   const categories = Singleton.value<Record<string, EmojiEntry[]>>();
   const all = Singleton.value<EmojiEntry[]>();
 
-  const emojiImagesUrl = Options.getEmotionsImageUrl(editor);
+  const emojiImagesUrl = Options.getEmojiImageUrl(editor);
 
   const getEmoji = (lib: RawEmojiEntry) => {
     // Note: This is a little hacky, but the database doesn't provide a way for us to tell what sort of database is being used
@@ -94,11 +94,11 @@ const initDatabase = (editor: Editor, databaseUrl: string, databaseId: string): 
 
   editor.on('init', () => {
     Resource.load(databaseId, databaseUrl).then((emojis) => {
-      const userEmojis = getUserDefinedEmoticons(editor);
+      const userEmojis = getUserDefinedEmoji(editor);
       processEmojis(Merger.merge(emojis, userEmojis));
     }, (err) => {
       // eslint-disable-next-line no-console
-      console.log(`Failed to load emoticons: ${err}`);
+      console.log(`Failed to load emojis: ${err}`);
       categories.set({});
       all.set([]);
     });